Add MPTCP support with the --multipath flag
Also available with the short option '-m'.
The MPTCP protocol is really a TCP variant, so this change
does not implement a new 'struct protocol'. Instead it just
extend the TCP support to optionally enable multipath.

The only required dependency is IPPROTO_MPTCP definition,
which should be provided by the netinet/in.h header.
To keep things simple, just conditionally provide the required
protocol, if the system header does not have it yet

pgagne: For cherry-pick, remove "idle-timeout" and "rcv-timeout" options
which are supported in this branch and where causing conflicts.